+++ packages/painter/src/types/common.ts
@@ -1,0 +1,10 @@
+// For CanvasRenderingContext2D
+export interface TextStyles {
+  font: string;
+  color: string;
+  textAlign: CanvasTextAlign;
+  textBaseline: CanvasTextBaseline;
+  direction: CanvasDirection;
+  lineHeight: number;    // px
+  letterSpacing: number; // px
+}
