+++ tsconfig.base.json
@@ -15,6 +15,11 @@
     "strict": true,
     "noUnusedLocals": true,
     "noUnusedParameters": true,
-    "noFallthroughCasesInSwitch": true
+    "noFallthroughCasesInSwitch": true,
+
+    "baseUrl": ".",
+    "paths": {
+      "@mirage-engine/painter": ["packages/painter/src/index.ts"]
+    }
   }
 }
